requirements.txt
Django==1.8.4
image==1.4.0
jieba==0.37
numpy==1.9.2
Pillow==2.9.0
wheel==0.24.0
